Cannon Tower, Currency, Camera, Outside border
The cannon tower was implemented, including the design of the tower and cannon ball, including the splash damage.

A simple currency system was coded which gradually increases. It also decreases based on the cost of each tower when one is placed.

Lateral and vertical camera movement was implemented

Outside border will now be automatically instantiated on the start of each level, regardless of its size or shape